rename pages and remove unnecessary folders
