+++ src/pages/narmesteleder/nyNLRequestAltinn.tsx
@@ -2,9 +2,9 @@ import { useState } from 'react';
 import { Alert, BodyShort, Loader } from '@navikt/ds-react';
 import { logger } from '@navikt/next-logger';
 
-import { withAuthenticatedPage } from '../../../auth/withAuth';
-import Innhold from '../../../components/Innhold/Innhold';
-import NyNLRequestAltinnForm from '../../../components/NyNLRequestAltinnForm/NyNLRequestAltinnForm';
+import { withAuthenticatedPage } from '../../auth/withAuth';
+import Innhold from '../../components/Innhold/Innhold';
+import NyNLRequestAltinnForm from '../../components/NyNLRequestAltinnForm/NyNLRequestAltinnForm';
 
 const NARMESTELEDER_URL = `/api/proxy/api/narmesteleder/request`;
